pyproject.toml
[project]
name = "neuro-desktop"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "aiohappyeyeballs==2.6.1",
    "aiohttp==3.13.1",
    "aiosignal==1.4.0",
    "annotated-doc==0.0.3",
    "annotated-types==0.7.0",
    "anyio==4.11.0",
    "async-timeout==5.0.1",
    "asyncio==4.0.0",
    "attrs==25.3.0",
    "certifi==2025.10.5",
    "cffi==2.0.0",
    "charset-normalizer==3.4.4",
    "click==8.3.0",
    "colorama==0.4.6",
    "cryptography==46.0.3",
    "easyocr==1.7.2",
    "exceptiongroup==1.3.0",
    "faker==37.5.3",
    "fastapi==0.120.0",
    "filelock==3.20.0",
    "flask>=3.1.2",
    "frozenlist==1.8.0",
    "fsspec==2025.9.0",
    "h11==0.16.0",
    "idna==3.11",
    "imageio==2.37.0",
    "importlib==1.0.4",
    "iniconfig==2.3.0",
    "jinja2==3.1.6",
    "jsf==0.11.2",
    "jsonschema==4.25.1",
    "jsonschema-specifications==2025.9.1",
    "lazy-loader==0.4",
    "libcomponent==0.0.5",
    "markupsafe==3.0.3",
    "mouseinfo==0.1.3",
    "mpmath==1.3.0",
    "multidict==6.7.0",
    "mypy-extensions==1.1.0",
    "nakuritycore==0.1.5",
    "networkx==3.4.2",
    "neuro-api==2.2.1",
    "neuro-api-tony==1.6.3",
    "ninja==1.13.0",
    "numpy==2.2.6",
    "opencv-python==4.12.0.88",
    "opencv-python-headless==4.12.0.88",
    "orjson==3.11.3",
    "outcome==1.3.0.post0",
    "packaging==25.0",
    "pathlib==1.0.1",
    "pillow==12.0.0",
    "pluggy==1.6.0",
    "propcache==0.4.1",
    "psutil==7.1.0",
    "pyautogui==0.9.54",
    "pyclipper==1.3.0.post6",
    "pycparser==2.23",
    "pydantic==2.12.3",
    "pydantic-core==2.41.4",
    "pygetwindow==0.0.9",
    "pygments==2.19.2",
    "pymsgbox==2.0.1",
    "pyperclip==1.11.0",
    "pyrect==0.2.0",
    "pyscreeze==1.0.1",
    "pytesseract==0.3.13",
    "pytest==8.4.2",
    "python-bidi==0.6.6",
    "pytweening==1.2.0",
    "pywin32==311 ; sys_platform == 'win32'",
    "pyyaml==6.0.3",
    "referencing==0.37.0",
    "requests==2.32.5",
    "rpds-py==0.28.0",
    "rstr==3.2.2",
    "scikit-image==0.25.2",
    "scipy==1.15.3",
    "semver==3.0.4",
    "shapely==2.1.2",
    "smart-open==7.4.1",
    "sniffio==1.3.1",
    "sortedcontainers==2.4.0",
    "starlette==0.48.0",
    "sympy==1.14.0",
    "tesseract==0.1.3",
    "tifffile==2025.5.10",
    "tomli==2.3.0",
    "torch==2.9.0",
    "torchvision==0.24.0",
    "trio==0.31.0",
    "trio-websocket==0.12.2",
    "typing-extensions==4.15.0",
    "typing-inspection==0.4.2",
    "tzdata==2025.2",
    "urllib3==2.5.0",
    "uv==0.9.3",
    "uvicorn==0.38.0",
    "websockets==15.0.1",
    "wrapt==2.0.0",
    "wsproto==1.2.0",
    "wxpython==4.2.3",
    "yarl==1.22.0",
]
